requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.4
django-bootstrap4==2.3.1
djangorestframework==3.12.2
importlib-metadata==3.1.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
soupsieve==2.0.1
sqlparse==0.4.1
zipp==3.4.0
